Fix the content handling of the input form
Fix the content handling when data is passed via the input form. The
previous handling prevented e.g. curl links by consuming the request
body. Now the request body is only consumed in case of the POST form.

Add also a test to check the `curl` handling of the pasta server.